Changed order of questions and updated to Christmas theme
