+++ code/src/App.css
@@ -1,8 +1,8 @@
 body {
   margin: 10px;
-  color: teal;
+  color: darkred;
   font-size: 18px;
-  background-color: #bef8e4;
+  background-color: #fffaf8;
 }
 
 form{
@@ -12,13 +12,28 @@ form{
 }
 input{
   margin: 15px;
-  border: 1px solid teal;
+  border: 1px solid palevioletred;
   font-size: 14px;
+  border-radius: 4px;
 }
 select{
   margin: 0 15px;
-  border: 1px solid teal;
+  border: 1px solid palevioletred;
   font-size: 14px;
 }
+button{
+width: 75px;
+font-size: 16px;
+font-weight: bold;
+background-color: crimson;
+color: white;
+padding: 5px;
+border: 1px solid black;
+border-radius: 5px;
+}
 .summary{
+}
+
+hr{
+  border-bottom: 5px solid palevioletred;
 }
